+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Make `ExtensionRegistry` iterable.
 * Remove `all()` method from `ExtensionRegistry`.
 * Add `get(**kwargs)` method to `ExtensionRegistry`.
+* Add package-specific exceptions.
 
 ## 0.0.1 (2018-06-11)